% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotPrjNetworks.R
\name{Date}
\alias{Date}
\title{Create date type vectors}
\usage{
Date(length = 0)
}
\arguments{
\item{length}{Length of the date type vector}
}
\value{
A date type vector of the indicated length
}
\description{
This function allows you to create date type vectors.
}
\examples{
Date(3)
Date(length=5)
}
\author{
Javier Celigueta Munoz
}

